# fix: `addMod` and `addMod8`