--- a/leopard.go
+++ b/leopard.go
@@ -841,7 +841,12 @@ func mulLog(a, log_b ffe) ffe {
 // z = x + y (mod kModulus)
 func addMod(a, b ffe) ffe {
 	sum := uint(a) + uint(b)
-	return ffe(sum % uint(modulus))
+	if sum == 65535 || sum == 131070 {
+		return 0
+	}
+
+	// Partial reduction step, allowing for kModulus to be returned
+	return ffe(sum + sum>>bitwidth)
 }
 
 // z = x - y (mod kModulus)