--- a/leopard8_test.go
+++ b/leopard8_test.go
@@ -15,7 +15,7 @@ func TestAddMod8(t *testing.T) {
 		{x: ffe8(254), y: ffe8(2), expected: ffe8(1)},
 		{x: ffe8(255), y: ffe8(0), expected: ffe8(0)},
 		{x: ffe8(255), y: ffe8(1), expected: ffe8(1)},
-		// {x: ffe8(255), y: ffe8(255), expected: ffe8(0)},
+		{x: ffe8(255), y: ffe8(255), expected: ffe8(0)},
 	}
 	for _, tc := range testCases {
 		got := addMod8(tc.x, tc.y)