Fixes an issue where `addMod` and `addMod8` would not perform modulus correctly.