@rootulp 2025-05-02
> Would it be possible to add a test where the wrong code manifests itself in data corruption?

Does "the wrong code" refer to the implementation on `main`? If yes, then these test cases already cover that. For example the test case

```
		{x: ffe(65534), y: ffe(1), expected: ffe(0)}, // fails on main because this returns 65535
```