@klauspost 2025-05-03
I would like to have a repro using public available functions, not just internal functions.

If it is only something that is triggered on specific setups, I would look into having an alternative implementation for that, since `return ffe(sum % uint(modulus))` is much slower than the current code.


This is the performance degradation, I would like to avoid... before/after.

```
BenchmarkReconstruct800x200/64-32 	    2466	    485145 ns/op	 131.92 MB/s	   49385 B/op	       5 allocs/op
BenchmarkReconstruct800x200/64-32 	    1896	    623928 ns/op	 102.58 MB/s	   49421 B/op	       5 allocs/op

BenchmarkReconstruct800x200/256-32         	    2167	    552150 ns/op	 463.64 MB/s	   49763 B/op	       5 allocs/op
BenchmarkReconstruct800x200/256-32         	    1710	    701377 ns/op	 365.00 MB/s	   49903 B/op	       5 allocs/op

BenchmarkReconstruct800x200/1024-32        	    1330	    898803 ns/op	1139.29 MB/s	   50836 B/op	       4 allocs/op
BenchmarkReconstruct800x200/1024-32        	    1148	   1035894 ns/op	 988.52 MB/s	   52969 B/op	       7 allocs/op

BenchmarkReconstruct800x200/4096-32        	     565	   2146278 ns/op	1908.42 MB/s	   64204 B/op	       7 allocs/op
BenchmarkReconstruct800x200/4096-32        	     525	   2268317 ns/op	1805.74 MB/s	   65334 B/op	       7 allocs/op

BenchmarkReconstruct800x200/16384-32       	     129	   9266695 ns/op	1768.05 MB/s	  309712 B/op	      19 allocs/op
BenchmarkReconstruct800x200/16384-32       	     127	   9388279 ns/op	1745.15 MB/s	  313831 B/op	      20 allocs/op
```