@rootulp 2025-05-06
I couldn't create a repro that uses external functions. `addMod` is used by many internal functions (`mulLog`, `fwht2`, `initFFTSkew`) but I don't have a clear way to repro `addMod` using just external functions.

If the performance degradation is the primary concern, I can iterate on the implementation to minimize the performance degradation. 

## Before
```
BenchmarkReconstructData800x200/64-12 	    1528	    757137 ns/op	  84.53 MB/s	   49343 B/op	       4 allocs/op
BenchmarkReconstructData800x200/256-12         	     823	   1463354 ns/op	 174.94 MB/s	   49922 B/op	       6 allocs/op
BenchmarkReconstructData800x200/1024-12        	     278	   4330686 ns/op	 236.45 MB/s	   56945 B/op	      10 allocs/op
BenchmarkReconstructData800x200/4096-12        	      70	  16287569 ns/op	 251.48 MB/s	  169770 B/op	      32 allocs/op
BenchmarkReconstructData800x200/16384-12       	      24	  61814656 ns/op	 265.05 MB/s	 1449397 B/op	      89 allocs/op
BenchmarkReconstructData800x200/65536-12       	       4	 275327000 ns/op	 238.03 MB/s	33616346 B/op	     514 allocs/op
BenchmarkReconstructData800x200/262144-12      	       2	 944879354 ns/op	 277.44 MB/s	268510252 B/op	    1031 allocs/op
BenchmarkReconstructData800x200/1048576-12     	       1	4824609250 ns/op	 217.34 MB/s	2147583872 B/op	    2054 allocs/op
```

## After
```
BenchmarkReconstruct800x200/64-12 	    1414	    836651 ns/op	  76.50 MB/s	   49355 B/op	       5 allocs/op
BenchmarkReconstruct800x200/256-12         	     765	   1569779 ns/op	 163.08 MB/s	   49976 B/op	       6 allocs/op
BenchmarkReconstruct800x200/1024-12        	     267	   4527271 ns/op	 226.18 MB/s	   57263 B/op	      11 allocs/op
BenchmarkReconstruct800x200/4096-12        	      66	  16325054 ns/op	 250.90 MB/s	  177081 B/op	      34 allocs/op
BenchmarkReconstruct800x200/16384-12       	      19	  66699502 ns/op	 245.64 MB/s	 1817836 B/op	     111 allocs/op
BenchmarkReconstruct800x200/65536-12       	       4	 289812594 ns/op	 226.13 MB/s	33616346 B/op	     514 allocs/op
BenchmarkReconstruct800x200/262144-12      	       1	1105943917 ns/op	 237.03 MB/s	536971040 B/op	    2053 allocs/op
BenchmarkReconstruct800x200/1048576-12     	       1	4785376458 ns/op	 219.12 MB/s	2147583872 B/op	    2054 allocs/op
```