@klauspost 2025-05-07
@rootulp The reason I am asking for the repro is that I have an (optimistic) hunch that the library allows for this value to be returned and compensates for it.

Maybe @catid is around or you could ask on the upstream repo: https://github.com/catid/leopard - since I suspect they would also like this fixed, if it is indeed a problem.