@liamsi 2025-05-12
I was able to hit the code path in question many times using public methods only (basically a encode/decode roundtrip) but without it causing any data corruption / integrity issues.
https://github.com/klauspost/reedsolomon/compare/master...liamsi:reedsolomon:tests-overfow

I'm still investigating but it's unclear how / if this can be misused at all.