@liamsi 2025-05-12
Actually, @catid explains why this isn't an issue: 
if addMod8 or addMod returns the modulus, these lines make sure we are still good:
- https://github.com/klauspost/reedsolomon/blob/1465e13be2461916127a8d89d9affa3e5a499ef1/leopard.go#L988
- https://github.com/klauspost/reedsolomon/blob/1465e13be2461916127a8d89d9affa3e5a499ef1/leopard8.go#L1074

See also the comment about this in the orig implemtation:
https://github.com/catid/leopard/blob/6e5725ebdf9da4370b0bcc4f70fa8eb66f4e6198/LeopardFF8.cpp#L192-L193